+++ lib/spec_forge/attribute/parameterized.rb
@@ -63,14 +63,6 @@ module SpecForge
       def prepare_arguments!
         @arguments = Attribute.from(arguments)
       end
-
-      def uses_positional_arguments?(method)
-        method.parameters.any? { |a| [:req, :opt, :rest].include?(a.first) }
-      end
-
-      def uses_keyword_arguments?(method)
-        method.parameters.any? { |a| [:keyreq, :key, :keyrest].include?(a.first) }
-      end
     end
   end
 end