+++ lib/spec_forge/configuration.rb
@@ -30,9 +30,6 @@ module SpecForge
     def initialize
       config = Normalizer.default_configuration
 
-      # Allows me to modify the error backtrace reporting within rspec
-      RSpec.configuration.instance_variable_set(:@backtrace_formatter, BacktraceFormatter)
-
       config[:base_url] = "http://localhost:3000"
       config[:factories] = Factories.new
       config[:specs] = RSpec.configuration